feat: :sparkles: Magento Products can have many Configurable Product Option Types
